ARM64 issues

+++ Classes/BattleClinicAPI/BCEveLoadout.h
@@ -10,16 +10,16 @@
 #import "BCRequest.h"
 
 @interface BCEveLoadout : BCRequest
-@property (nonatomic, assign) NSInteger loadoutID;
-@property (nonatomic, assign) NSInteger typeID;
-@property (nonatomic, assign) NSInteger thumbsUp;
-@property (nonatomic, assign) NSInteger thumbsDown;
-@property (nonatomic, assign) NSInteger topic;
+@property (nonatomic, assign) int32_t loadoutID;
+@property (nonatomic, assign) int32_t typeID;
+@property (nonatomic, assign) int32_t thumbsUp;
+@property (nonatomic, assign) int32_t thumbsDown;
+@property (nonatomic, assign) int32_t topic;
 @property (nonatomic, copy) NSString *title;
 @property (nonatomic, copy) NSString *description;
 @property (nonatomic, copy) NSString *fitting;
 
-+ (id) eveLoadoutsWithAPIKey:(NSString*) apiKey loadoutID:(NSInteger)loadoutID error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
-- (id) initWithAPIKey:(NSString*) apiKey loadoutID:(NSInteger)loadoutID error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
++ (id) eveLoadoutsWithAPIKey:(NSString*) apiKey loadoutID:(int32_t)loadoutID error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
+- (id) initWithAPIKey:(NSString*) apiKey loadoutID:(int32_t)loadoutID error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
 
 @end
