+++ Classes/EVECentralAPI/EVECentralMarketStat.h
@@ -24,7 +24,7 @@
 
 
 @interface EVECentralMarketStatType: NSObject<NSCoding>
-@property (nonatomic) NSInteger typeID;
+@property (nonatomic) int32_t typeID;
 @property (nonatomic, strong) EVECentralMarketStatTypeStat *all;
 @property (nonatomic, strong) EVECentralMarketStatTypeStat *buy;
 @property (nonatomic, strong) EVECentralMarketStatTypeStat *sell;
@@ -37,6 +37,6 @@
 @interface EVECentralMarketStat : EVECentralRequest
 @property (nonatomic, strong) NSArray *types;
 
-+ (id) marketStatWithTypeIDs: (NSArray*) typeIDs regionIDs: (NSArray*) regionIDs hours: (NSInteger) hours minQ: (NSInteger) minQ cachePolicy:(NSURLRequestCachePolicy) cachePolicy error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
-- (id) initWithTypeIDs: (NSArray*) typeIDs regionIDs: (NSArray*) regionIDs hours: (NSInteger) hours minQ: (NSInteger) minQ cachePolicy:(NSURLRequestCachePolicy) cachePolicy error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
++ (id) marketStatWithTypeIDs: (NSArray*) typeIDs regionIDs: (NSArray*) regionIDs hours: (int32_t) hours minQ: (int32_t) minQ cachePolicy:(NSURLRequestCachePolicy) cachePolicy error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
+- (id) initWithTypeIDs: (NSArray*) typeIDs regionIDs: (NSArray*) regionIDs hours: (int32_t) hours minQ: (int32_t) minQ cachePolicy:(NSURLRequestCachePolicy) cachePolicy error:(NSError **)errorPtr progressHandler:(void(^)(CGFloat progress, BOOL* stop)) progressHandler;
 @end
