+++ Classes/EVEDBAPI/EVEDBInvBlueprintType.h
@@ -10,30 +10,30 @@
 
 @class EVEDBInvType;
 @interface EVEDBInvBlueprintType : EVEDBObject
-@property (nonatomic) NSInteger blueprintTypeID;
+@property (nonatomic) int32_t blueprintTypeID;
 @property (nonatomic, strong) EVEDBInvType* blueprintType;
-@property (nonatomic) NSInteger parentBlueprintTypeID;
+@property (nonatomic) int32_t parentBlueprintTypeID;
 @property (nonatomic, strong) EVEDBInvBlueprintType* parentBlueprintType;
-@property (nonatomic) NSInteger productTypeID;
+@property (nonatomic) int32_t productTypeID;
 @property (nonatomic, strong) EVEDBInvType* productType;
-@property (nonatomic) NSInteger productionTime;
-@property (nonatomic) NSInteger techLevel;
-@property (nonatomic) NSInteger researchProductivityTime;
-@property (nonatomic) NSInteger researchMaterialTime;
-@property (nonatomic) NSInteger researchCopyTime;
-@property (nonatomic) NSInteger researchTechTime;
-@property (nonatomic) NSInteger productivityModifier;
-@property (nonatomic) NSInteger materialModifier;
-@property (nonatomic) NSInteger wasteFactor;
-@property (nonatomic) NSInteger maxProductionLimit;
+@property (nonatomic) int32_t productionTime;
+@property (nonatomic) int32_t techLevel;
+@property (nonatomic) int32_t researchProductivityTime;
+@property (nonatomic) int32_t researchMaterialTime;
+@property (nonatomic) int32_t researchCopyTime;
+@property (nonatomic) int32_t researchTechTime;
+@property (nonatomic) int32_t productivityModifier;
+@property (nonatomic) int32_t materialModifier;
+@property (nonatomic) int32_t wasteFactor;
+@property (nonatomic) int32_t maxProductionLimit;
 @property (nonatomic, strong) NSMutableDictionary* ramTypeRequirements;
 @property (nonatomic, strong) NSArray *typeMaterials;
 
 
-+ (id) invBlueprintTypeWithBlueprintTypeID:(NSInteger)aBlueprintTypeID error:(NSError **)errorPtr;
-- (id) initWithBlueprintTypeID:(NSInteger)aBlueprintTypeID error:(NSError **)errorPtr;
++ (id) invBlueprintTypeWithBlueprintTypeID:(int32_t)aBlueprintTypeID error:(NSError **)errorPtr;
+- (id) initWithBlueprintTypeID:(int32_t)aBlueprintTypeID error:(NSError **)errorPtr;
 
 - (NSArray*) activities;
-- (NSArray*) requiredSkillsForActivity:(NSInteger) activityID;
-- (NSArray*) requiredMaterialsForActivity:(NSInteger) activityID;
+- (NSArray*) requiredSkillsForActivity:(int32_t) activityID;
+- (NSArray*) requiredMaterialsForActivity:(int32_t) activityID;
 @end
